requirements.txt
ipykernel
langchain
pydantic 
python-dotenv
langchain-community
pypdf 
bs4
arxiv
pymupdf
rich 
langchain-text-splitters
langchain-openai
chromadb
langchain-groq
beautifulsoup4
faiss-cpu
streamlit